server/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.3
cryptography==3.4.7
defusedxml==0.7.1
DateTime==4.3
Deprecated==1.2.12
Django==3.2
django-environ==0.4.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
environ==1.0
idna==3.2
oauthlib==3.1.0
lxml==4.6.3
numpy==1.21.1
pandas==1.3.0
Pillow==8.3.1
pykrx==1.0.17
python-dateutil==2.8.2
pytz==2021.1
pycparser==2.20
PyJWT==1.7.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
wincertstore==0.2
wrapt==1.12.1
xlrd==2.0.1
zope.interface==5.4.0